Started implementing DecodableDataProviderResponse
