+++ Tests/HTTPLibTests/HTTPDataProviderTests.swift
@@ -140,81 +140,6 @@ extension HTTPDataProviderTests {
 }
 
 extension HTTPDataProviderTests {
-	func testDataProviderResponseMappedResponseError() throws {
-		struct MockError: Swift.Error { }
-		XCTAssertEqual(
-			DataProviderResponse.mappedResponse(
-				from: nil,
-				response: nil,
-				error: MockError()),
-			.unreachable)
-	}
-
-	func testDataProviderResponseMappedResponseNotHTTPResponse() throws {
-		XCTAssertEqual(
-			DataProviderResponse.mappedResponse(
-				from: "some-data".data(using: .utf8)!,
-				response: URLResponse(
-					url: .init(string: "test-url")!,
-					mimeType: nil,
-					expectedContentLength: 0,
-					textEncodingName: nil),
-				error: nil),
-			.error(
-				code: nil,
-				data: "some-data".data(using: .utf8)!)
-		)
-	}
-
-	func testDataProviderResponseMappedResponseHTTPErrorNoStatusCode() throws {
-		XCTAssertEqual(
-			DataProviderResponse.mappedResponse(
-				from: "some-data".data(using: .utf8)!,
-				response: HTTPURLResponse(
-					url: .init(string: "test-url")!,
-					statusCode: 0,
-					httpVersion: nil,
-					headerFields: nil),
-				error: nil),
-			.error(
-				code: nil,
-				data: "some-data".data(using: .utf8)!)
-		)
-	}
-
-	func testDataProviderResponseMappedResponseHTTPError() throws {
-		XCTAssertEqual(
-			DataProviderResponse.mappedResponse(
-				from: "some-data".data(using: .utf8)!,
-				response: HTTPURLResponse(
-					url: .init(string: "test-url")!,
-					statusCode: 404,
-					httpVersion: nil,
-					headerFields: nil),
-				error: nil),
-			.error(
-				code: .notFound,
-				data: "some-data".data(using: .utf8)!)
-		)
-	}
-
-	func testDataProviderResponseMappedResponseSuccess() throws {
-		XCTAssertEqual(
-			DataProviderResponse.mappedResponse(
-				from: "some-data".data(using: .utf8)!,
-				response: HTTPURLResponse(
-					url: .init(string: "test-url")!,
-					statusCode: 200,
-					httpVersion: nil,
-					headerFields: nil),
-				error: nil),
-			.success(
-				data: "some-data".data(using: .utf8)!)
-		)
-	}
-}
-
-extension HTTPDataProviderTests {
 	func testURLRequestFromDataProviderRequestSuccess() throws {
 		let request = RawDataProviderRequest(
 			method: .get,
